requirements.txt
Django
mysqlclient
django-dotenv
django-extensions
jupyter
python-magic
channels
djangorestframework
markdown
django-filter